+++ borsh-derive/src/lib.rs
@@ -4,12 +4,11 @@ use proc_macro::TokenStream;
 use proc_macro2::Span;
 use proc_macro_crate::crate_name;
 use proc_macro_crate::FoundCrate;
-use quote::ToTokens;
 use syn::{Ident, ItemEnum, ItemStruct, ItemUnion};
 
+use borsh_derive_internal::check_use_discriminant;
 #[cfg(feature = "schema")]
 use borsh_schema_derive_internal::{process_enum, process_struct};
-use syn::{parse2, Meta, MetaNameValue};
 use syn::{parse_macro_input, DeriveInput};
 
 #[proc_macro_derive(BorshSerialize, attributes(borsh_skip, borsh))]
@@ -78,56 +77,6 @@ pub fn borsh_deserialize(input: TokenStream) -> TokenStream {
     })
 }
 
-fn check_use_discriminant(derive_input: DeriveInput) -> Result<Option<bool>, TokenStream> {
-    for attr in &derive_input.attrs {
-        if attr.path().is_ident("borsh") {
-            if let Meta::List(list) = attr.meta.clone() {
-                let tokens = list.tokens;
-                let meta: Meta = parse2(tokens).map_err(|err| err.to_compile_error())?;
-
-                if let Meta::NameValue(value) = meta {
-                    let MetaNameValue { path, value, .. } = value;
-                    if !path.is_ident("use_discriminant") {
-                        return Err(TokenStream::from(
-                            syn::Error::new(
-                                derive_input.ident.span(),
-                                "`use_discriminant` is the only supported attribute for `borsh`",
-                            )
-                            .to_compile_error(),
-                        ));
-                    }
-
-                    if let syn::Data::Struct(ref _data) = derive_input.data {
-                        return Err(TokenStream::from(
-                            syn::Error::new(
-                                derive_input.ident.span(),
-                                "borsh (use_discriminant=<bool>) does not support structs",
-                            )
-                            .to_compile_error(),
-                        ));
-                    }
-
-                    let value = value.to_token_stream().to_string();
-                    return match value.as_str() {
-                        "true" => Ok(Some(true)),
-                        "false" => Ok(Some(false)),
-                        _ => {
-                            return Err(TokenStream::from(
-                                syn::Error::new(
-                                    derive_input.ident.span(),
-                                    "`use_discriminant` accept only `true` or `false`",
-                                )
-                                .to_compile_error(),
-                            ));
-                        }
-                    };
-                }
-            }
-        }
-    }
-    Ok(None)
-}
-
 #[cfg(feature = "schema")]
 #[proc_macro_derive(BorshSchema, attributes(borsh_skip, use_discriminant))]
 pub fn borsh_schema(input: TokenStream) -> TokenStream {
@@ -155,73 +104,4 @@ pub fn borsh_schema(input: TokenStream) -> TokenStream {
         Ok(res) => res,
         Err(err) => err.to_compile_error(),
     })
-}
-
-#[cfg(test)]
-mod tests {
-
-    use super::*;
-    use quote::quote;
-    #[test]
-    fn test_check_use_discriminant() {
-        let item_enum: DeriveInput = syn::parse2(quote! {
-            #[derive(BorshDeserialize, Debug)]
-            #[borsh(use_discriminant = false)]
-            enum AWithUseDiscriminantFalse {
-                X,
-                Y,
-            }
-        })
-        .unwrap();
-        let actual = check_use_discriminant(item_enum).unwrap();
-
-        assert_eq!(actual, Some(false));
-    }
-
-    #[test]
-    fn test_check_use_discriminant_true() {
-        let item_enum: DeriveInput = syn::parse2(quote! {
-            #[derive(BorshDeserialize, Debug)]
-            #[borsh(use_discriminant = true)]
-            enum AWithUseDiscriminantFalse {
-                X,
-                Y,
-            }
-        })
-        .unwrap();
-        let actual = check_use_discriminant(item_enum).unwrap();
-
-        assert_eq!(actual, Some(true));
-    }
-
-    #[test]
-    #[should_panic]
-    fn test_check_use_discriminant_wrong_value() {
-        let item_enum: DeriveInput = syn::parse2(quote! {
-            #[derive(BorshDeserialize, Debug)]
-            #[borsh(use_discriminant = 111)]
-            enum AWithUseDiscriminantFalse {
-                X,
-                Y,
-            }
-        })
-        .unwrap();
-        let actual = check_use_discriminant(item_enum);
-        assert!(actual.is_err());
-    }
-    #[test]
-    #[should_panic]
-    fn test_check_use_discriminant_on_struct() {
-        let item_enum: DeriveInput = syn::parse2(quote! {
-            #[derive(BorshDeserialize, Debug)]
-            #[borsh(use_discriminant = false)]
-            struct AWithUseDiscriminantFalse {
-                X,
-                Y,
-            }
-        })
-        .unwrap();
-        let actual = check_use_discriminant(item_enum);
-        assert!(actual.is_err());
-    }
 }
